backend/requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
asttokens==3.0.1
async-lru==2.1.0
async-timeout==4.0.3
attrs==25.4.0
babel==2.18.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.3.1
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
effdet==0.4.1
email-validator==2.3.0
emoji==2.15.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.8
fastapi-cli==0.0.21
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastjsonschema==2.21.2
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-vision==3.12.1
google-genai==1.63.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==7.2.0
ipython==8.38.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==0.3.14
langchain-chroma==0.2.2
langchain-community==0.3.14
langchain-core==0.3.30
langchain-google-genai==2.0.9
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langdetect==1.0.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-sdk==0.3.5
langsmith==0.2.11
lark==1.3.1
layoutparser==0.3.4
llvmlite==0.46.0
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
msoffcrypto-tool==6.0.0
multidict==6.7.1
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
notebook==7.5.3
notebook_shim==0.2.4
numba==0.63.1
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
openai==1.109.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pi_heif==1.2.0
pikepdf==10.3.0
pillow==12.1.1
platformdirs==4.5.1
portalocker==3.2.0
posthog==5.4.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pypandoc==1.16.2
pyparsing==3.3.2
pypdf==6.7.0
pypdfium2==5.4.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-json-logger==4.0.0
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.2
rich-toolkit==0.19.4
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.15.3
Send2Trash==2.1.0
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
terminado==0.18.1
tiktoken==0.12.0
timm==1.0.24
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.1.0
typer==0.23.0
typer-slim==0.23.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.32
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
python-jose==3.5.0
PyJWT==2.11.0
ecdsa==0.19.1
